+++ launcher/LauncherIcon.cpp
@@ -1128,7 +1128,7 @@ LauncherIcon::OnRemoteCountChanged(LauncherEntryRemote* remote)
   if (remote->Count() > 9999)
     text = "****";
   else
-    text = std::to_string(remote->Count());
+    text = std::to_string((long long)remote->Count());
 
   SetEmblemText(text);
 }
@@ -1162,7 +1162,7 @@ LauncherIcon::OnRemoteCountVisibleChanged(LauncherEntryRemote* remote)
 {
   if (remote->CountVisible())
   {
-    SetEmblemText(std::to_string(remote->Count()));
+    SetEmblemText(std::to_string((long long)remote->Count()));
   }
   else
   {
